Fall back to localhost copying of artifacts for podman connections
Podman is not a supported connection for the `synchronize` module so it
fails to copy the artifacts. Falling back to localhost artifacts code path
for the artifacts mounted to the container.

```
synchronize uses rsync to function. rsync needs to connect to the remote
host via ssh, docker client or a direct filesystem copy. This remote
host is being accessed via containers.podman.podman instead so it cannot
work.
```